+++ CHANGELOG.md
@@ -19,6 +19,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com), and this 
 - Background colours of the switch inputs.
 - Colours of the images in the cover components to the new brand colours.
 - Colours of the images in the rule components to the new brand colours.
+- Layout of the authors page.
 
 ### Fixed